requirements.txt
python-dotenv>=1.0.0

# Database + drivers
SQLAlchemy>=2.0
psycopg2-binary>=2.9
PyMySQL>=1.1

# LangChain (LLM + SQL agent)
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0

# Desktop UI
PySide6>=6.6.0
langsmith>=0.1.70


